183 host add webusb (#184)
- Added configurable settings
- Added player detection capability 
   - By default, poll each player. Once detection is made, permanently enable that controller.
   - Can be changed in settings 
   - This removes the need to have 1, 2, 4 player variants of the binary as this can be configured at runtime by the user
   - Player detection modes: `Disabled`, `Enabled`, `Disabled, Enable On Detect`, `Auto`, `Auto, No Disable`
- Added WebUSB interfaces, and all functionality may be accessed through this interface
- Disable Mass Storage Class (MSC) by default but allow this to be enabled through the WebUSB interface
    - I am not going to support any issues with MSC going forward
    - Read/write of VMUs may now be accessed through WebUSB and the webpage
- Added WebUSB page which can change settings: https://orangefox86.github.io/DreamPicoPort/webusb/
    - Safari and Firefox don't support WebUSB. As far as I can tell, only Chromium-based browsers support this. These browsers capture about 70% of users
    - Firefox backs up its decision not to support WebUSB due to security concerns, but the user must be asked to select a device before the page gets access to it (unless the page is running locally). I'm a Firefox user myself, so it sucks that I have to manually open in another browser, but I don't feel that will be a deal breaker for many.
- USB version now matches the release version
- Serial is now set in all USB string descriptors - this will provide a way to link an SDL instance to libUSB instance more easily on Linux and Mac since serial is not always exposed due to permissions, but the device name is
    - I would like to allow multiple instances of the hardware on a system to function in the future whereas right now, flycast only supports 1 DPP board plugged in
- Tweaks to Maple packet handling so that byte order may be specified (reduces the need to swap bytes on WebUSB and TTY interfaces)
- Allow the Maple packet handling to hold a `shared_ptr` to something. I'm leveraging this in the WinUSB implementation so that I can create an object and automatically have it destroyed once a response is sent over WinUSB due to Maple communication completion.
- Added XO command which will provide information about what gamepads are available and which are connected; only auto setting should cause disconnect

+++ src/hostLib/ScreenData.hpp
@@ -62,7 +62,7 @@ class ScreenData
 
     public:
         //! Number of words in a screen
-        static const uint32_t NUM_SCREEN_WORDS = 48;
+        static const uint8_t NUM_SCREEN_WORDS = 48;
         //! Number of default screens
         static const uint32_t NUM_DEFAULT_SCREENS = 4;
